Refactor primitives crate, have types in a submodule
