+++ runtime/manta/src/fee.rs
@@ -19,7 +19,7 @@ use frame_support::weights::{
 	constants::ExtrinsicBaseWeight, WeightToFeeCoefficient, WeightToFeeCoefficients,
 	WeightToFeePolynomial,
 };
-use manta_primitives::Balance;
+use manta_primitives::types::Balance;
 use smallvec::smallvec;
 pub use sp_runtime::Perbill;
