+++ runtime/primitives/src/constants.rs
@@ -26,7 +26,7 @@ pub const MANTA_TOKEN_SYMBOL: &str = "MANTA";
 
 /// Manta parachain time-related
 pub mod time {
-	use crate::{BlockNumber, Moment};
+	use crate::types::{BlockNumber, Moment};
 	/// This determines the average expected block time that we are targeting. Blocks will be
 	/// produced at a minimum duration defined by `SLOT_DURATION`. `SLOT_DURATION` is picked up by
 	/// `pallet_timestamp` which is in turn picked up by `pallet_aura` to implement `fn
